+++ packages/ai/src/agent/create-agent-ui-stream-response.ts
@@ -1,0 +1,41 @@
+import { UIMessageStreamOptions } from '../generate-text';
+import { ToolSet } from '../generate-text/tool-set';
+import { createUIMessageStreamResponse } from '../ui-message-stream';
+import { UIMessageStreamResponseInit } from '../ui-message-stream/ui-message-stream-response-init';
+import { InferUITools, UIMessage } from '../ui/ui-messages';
+import { Agent } from './agent';
+import { createAgentUIStream } from './create-agent-ui-stream';
+
+/**
+ * Runs the agent and returns a response object with a UI message stream.
+ *
+ * @param agent - The agent to run.
+ * @param messages - The input UI messages.
+ *
+ * @returns The response object.
+ */
+export async function createAgentUIStreamResponse<
+  TOOLS extends ToolSet = {},
+  OUTPUT = never,
+  OUTPUT_PARTIAL = never,
+>({
+  headers,
+  status,
+  statusText,
+  consumeSseStream,
+  ...options
+}: {
+  agent: Agent<TOOLS, OUTPUT, OUTPUT_PARTIAL>;
+  messages: unknown[];
+} & UIMessageStreamResponseInit &
+  UIMessageStreamOptions<
+    UIMessage<never, never, InferUITools<TOOLS>>
+  >): Promise<Response> {
+  return createUIMessageStreamResponse({
+    headers,
+    status,
+    statusText,
+    consumeSseStream,
+    stream: await createAgentUIStream(options),
+  });
+}